ADD: Rotas para verificar se a categoria, tipo e/ou caracteristicas estão com dependencia em outras tabelas e rotas para excluir elas caso não estejam dependentes

+++ model/controllers/controler_categorias.py
@@ -5,6 +5,34 @@ class Categoria:
 
 # CATEGORIA ------------------------------------------------------------------------------------------------------#
 
+    # Verifica se a Categoria está ligada a alguma Estante ou Produto
+    def verificar_dependencia_categoria(cod_categoria):
+
+        conexao = Conection.create_connection()
+
+        cursor = conexao.cursor()
+
+        # Verifica se a categoria está em alguma estante ou em algum produto
+        sql = """
+            SELECT EXISTS (
+                SELECT 1 FROM estante WHERE cod_categoria = %s
+                UNION ALL
+                SELECT 1 FROM produto WHERE cod_categoria = %s
+            ) AS dependencia;
+        """
+
+        valores = (cod_categoria, cod_categoria)
+        
+        # Executa a consulta
+        cursor.execute(sql, valores)
+        
+        # O resultado será (1,) se houver dependência, ou (0,) se não houver
+        dependencia = cursor.fetchone()[0] == 1
+
+        cursor.close()
+        conexao.close()
+        return dependencia # Retorna True se houver dependência
+
     # Conexao com o banco de dados para criar uma categoria
     def cadastrar_categoria(nome, cpf):
 
@@ -54,20 +82,54 @@ class Categoria:
     # Conexao com o banco de dados para excluir uma categoria
     def remover_categoria(cod_categoria):
 
-        conexao = Conection.create_connection()
+        # Verifica se a categoria possui uma dependencia 
+        if Categoria.verificar_dependencia_categoria(cod_categoria):
+            # Retorna se a remoção falhou por conta da dependência
+            return False # Não pode excluir
 
+        # Se não possuir uma dependencia, executa a exclusão da categoria
+        conexao = Conection.create_connection()
         cursor = conexao.cursor()
 
         sql = "DELETE FROM categoria WHERE cod_categoria = %s;"
 
-        cursor.execute(sql, (cod_categoria,))
+        valor = (cod_categoria,)
+
+        cursor.execute(sql, valor)
+
         conexao.commit()
         
         cursor.close()
         conexao.close()
-        return True 
+        return True
     
 # TIPO ------------------------------------------------------------------------------------------------------#
+
+    # Verifica se o Tipo está ligado a alguma Estante ou Produto
+    def verificar_dependencia_tipo(cod_tipo):
+
+        conexao = Conection.create_connection()
+
+        cursor = conexao.cursor()
+
+        # Verifica se o tipo está em alguma categoria
+        sql = """
+            SELECT EXISTS (
+                SELECT 1 FROM caracteristica WHERE cod_tipo = %s
+            ) AS dependencia;
+        """
+
+        valor = (cod_tipo,)
+        
+        # Executa a consulta
+        cursor.execute(sql, valor)
+        
+        # O resultado será (1,) se houver dependência, ou (0,) se não houver
+        dependencia = cursor.fetchone()[0] == 1
+
+        cursor.close()
+        conexao.close()
+        return dependencia # Retorna True se houver dependência
 
     # Conexao com o banco de dados para criar um tipo com base em uma categoria
     def cadastrar_tipo_categoria(nome, cpf, cod_categoria):
@@ -114,23 +176,57 @@ class Categoria:
         return resultado
     
     # Conexao com o banco de dados para excluir um tipo
-    def remover_tipo(cod_categoria):
+    def remover_tipo(cod_tipo):
+
+        # Verifica se a categoria possui uma dependencia 
+        if Categoria.verificar_dependencia_tipo(cod_tipo):
+            # Retorna se a remoção falhou por conta da dependência
+            return False # Não pode excluir
+
+        # Se não possuir uma dependencia, executa a exclusão do tipo
+        conexao = Conection.create_connection()
+        cursor = conexao.cursor()
+
+        sql = "DELETE FROM tipo WHERE cod_tipo = %s;"
+
+        valor = (cod_tipo,)
+
+        cursor.execute(sql, valor)
+
+        conexao.commit()
+        
+        cursor.close()
+        conexao.close()
+        return True
+    
+# CARACTERISTICA ------------------------------------------------------------------------------------------------------#
+
+   # Verifica se a Caracteristica está ligada a alguma Estante ou Produto
+    def verificar_dependencia_caracterisica(cod_caracteristica):
 
         conexao = Conection.create_connection()
 
         cursor = conexao.cursor()
 
-        sql = "DELETE FROM tipo WHERE cod_tipo = %s;"
+        # Verifica se a categoria está em alguma estante ou em algum produto
+        sql = """
+            SELECT EXISTS (
+                SELECT 1 FROM produto WHERE cod_caracteristica = %s
+            ) AS dependencia;
+        """
 
-        cursor.execute(sql, (cod_categoria,))
-        conexao.commit()
+        valor = (cod_caracteristica,)
         
+        # Executa a consulta
+        cursor.execute(sql, valor)
+        
+        # O resultado será (1,) se houver dependência, ou (0,) se não houver
+        dependencia = cursor.fetchone()[0] == 1
+
         cursor.close()
         conexao.close()
-        return True 
+        return dependencia # Retorna True se houver dependência
     
-# CARACTERISTICA ------------------------------------------------------------------------------------------------------#
-
     # Conexao com o banco de dados para criar uma caracteristica com base no tipo
     def cadastrar_tipo_caracteristica(nome, cod_tipo, cpf):
 
@@ -175,18 +271,28 @@ class Categoria:
 
         return resultado
     
-    # Conexao com o banco de dados para excluir uma caracteristica
+    
+        # Conexao com o banco de dados para excluir uma caracteristica
     def remover_caracteristica(cod_caracteristica):
 
-        conexao = Conection.create_connection()
+        # Verifica se a categoria possui uma dependencia 
+        if Categoria.verificar_dependencia_caracterisica(cod_caracteristica):
+            # Retorna se a remoção falhou por conta da dependência
+            return False # Não pode excluir
 
+        # Se não possuir uma dependencia, executa a exclusão do tipo
+        conexao = Conection.create_connection()
         cursor = conexao.cursor()
 
         sql = "DELETE FROM caracteristica WHERE cod_caracteristica = %s;"
 
-        cursor.execute(sql, (cod_caracteristica,))
+        valor = (cod_caracteristica,)
+
+        cursor.execute(sql, valor)
+
         conexao.commit()
         
         cursor.close()
         conexao.close()
-        return True 
+        return True
+
